[framework] refactor reduction algorithm
After merging a post-condition, root node folds all published channels
on it. If a channel is successufly captured by a reduction, we merge the
resulting post-condition, remove the used channel from the channel pool of
the next iteration and add the new channels of the reduction to the pool.
Otherwise, if a channel is not used, we keep it in the pool of the next
iteration.

When we are done with all channels, we check if the new pool is empty.
In this case, we return the current post-condition, otherwise we begin
a new iteration. The maximum number of iterations is bounded by the
command line option -reduce-iter.