[python] consider all Python values as objects
Atomic (int, float, bool, str, None, NotImplemented) are given fixed
addresses. Mapping between variables and addresses is maintained by the
domain memory.addr_env.

When a variable v is evaluated, if it is associated to an atomic type
address a, a symbolic expression E_py_addr_value(a, <v>) is returned
that embeds the value expression <v> associated to variable v.