[framework] Generation of the html eval/exec tree
This needs to be improved a lot:
* modifying the printer print_eb_ed print_sb_sd in ouput/html.ml should
allow displaying: zones, eval/exec results and stuff
* Interface of analyzer was modified because the types of the actions
printed in html depends upon the Domain.t. This is a temporary solution.

example of use:
./bin/mopsa-universal -html=./outputs/html/res.html
./examples/universal/ex2.u

This will generate a outputs/html/res.html file with foldable
expressions and statements (provided the css file is in the same
directory, which it is)