[c] cell: put old information in bases
The base of a cell indicates which version is accessed in the post-condition:
the old one (pre-condition) or the new one. Previously, this information was stored
in the cell itself: we had two kinds of cells. By moving it to the base level, only
one kind of cells is necessary.