[framework] disable temporarily -collect-states option
In order to implement the reduced product efficiently, the
non-relational domain should be implemented with the simplified domain
signature. However, the keeping reachable states is done using the
flow-insensitive context, which is not provided in the simplified
signature. So, this option is removed for the moment to test the reduced
product.