[universal] loops: fix merging of cur and continue flows
The function `merge_cur_and_continue` used `Flow.map` to merge `T_cur`
and `T_continue` flows. However, `Flow.map` iterates over existing
tokens, so when `T_cur` is not in the map (i.e. it is equal to bottom),
the result if `merge_cur_and_continue` is always bottom, even if
`T_continue` is present.