[*] distinguish between the mode of a variable and the mode of variable expression
Decouple the access mode into two layers:

- A mode is attached to the variable depending on its "concrete
cardinality", i.e. a strong variable represents a single concrete
variable and a weak variable represents multiple concrete ones.

- An optional mode can also be attached to a variable expression. If
present, it overwrites the mode of the variable. This is useful when one
wants to access a weak variable with strong updates for example.