[c] string length: return unchanged flow in quantified forget on scalar
When handled a quantified forget statement on a scalar base, there is no
need to return Cases.not_handled, because the scalar domain will not
handle it.

Note that the StringLength domain needs to be put in a product with
another domain that handles this statement (e.g. Cells). Otherwise, the
analysis is not sound.