[c] Create temporary variable to handle guard of switch statements.
I discovered the issue in the reduced testcase for ln:
```c
extern int optarg;

int numbered_backup_15462(int x) { return 2; }

backupfile_internal() {
  int dirp;
  switch (numbered_backup_15462(dirp))
  case 2:
    return 1;
}
do_link_347336() { backupfile_internal(); }

main() {
  getopt_long(1, 1, "", 1, 1);
  if (optarg)
    backupfile_internal();
  else
    do_link_347336();
}
```

For a reason I don't completely understand, the return variable of
`numbered_backup` in the switch guard is not cleaned until the end of
the whole switch. As the C analysis does not clean all flows, we have
remnants of this return variable. This is weird because an if statement
would handle things correctly and run the cleaners of its guard before
analyzing the statements it contains. The side approach I decided to use
was to create an auxiliary variable in the meantime.