[framework] partitioning: fix print_expr
When calling `man.print_expr` with a configuration containing a partitioning
domain, partitions were duplicated and placed at wrong places.

The bug is due to the fact that `print_expr` in the partitioning
combiner re-prints the partition when called recursively.

A solution is to add a global flag to avoid printing the partition when
calling `print_expr` of the partitioned domain.